src: standardize byte/packet counter members to use `long`
Application-level byte and packet counters do not need a fixed-width
integer type; using `long` is the established INET convention (see
TcpSinkApp, TcpEchoApp, TcpGenericServerApp, EtherTrafGen, etc.).
The previous types were inconsistent across the SCTP app classes:

  - SctpClient/SctpServer used `unsigned long int` (verbose, unsigned
    without good reason)
  - SctpPeer used plain `int` (too narrow for a byte counter)
  - SctpNatPeer/SctpNatServer mixed `int64_t` (bytesSent) with
    `int32_t` (bytesRcvd, packetsSent, ...) -- clearly inconsistent
  - TcpConnection/TcpLwipConnection used `int32_t` for
    maxByteCountRequested (a byte count)

All of the above are now uniformly `long`.  The per-association map
value types in SctpNatPeer (RcvdPacketsPerAssoc, SentPacketsPerAssoc,
RcvdBytesPerAssoc) are changed from `int64_t` to `long` for the same
reason.  The one `static_cast<unsigned long int>` in SctpServer.cc
that compared against a `long` field is updated to `(long)` to match.

+++ src/inet/applications/sctpapp/SctpNatServer.h
@@ -45,9 +45,9 @@ class INET_API SctpNatServer : public SimpleModule, public LifecycleUnsupported
     uint32_t assocId;
     SctpSocket *socket;
     bool shutdownReceived;
-    int64_t bytesSent;
-    int32_t packetsSent;
-    int32_t packetsRcvd;
+    long bytesSent;
+    long packetsSent;
+    long packetsRcvd;
     int32_t numSessions;
     int32_t numRequestsToSend; // requests to send in this session
     bool ordered;
